Transition indices to NSUInteger
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2599 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/DataBasin/DBSoapCSV.m
@@ -247,8 +247,8 @@
 
 - (void)describeSObject: (NSString *)objectType toWriter:(DBCVSWriter *)writer
 {
-  int            i;
-  int            size;
+  NSUInteger      i;
+  NSUInteger     size;
   DBSObject      *object;
   NSDictionary   *properties;
   NSArray        *fields;
@@ -271,7 +271,7 @@
   for (i = 0; i < size; i++)
     {
       NSMutableArray *values;
-      int            j;
+      NSUInteger      j;
       NSString       *field;
       
       field = [fields objectAtIndex: i];
